requirements.txt
Flask
celery
python-ldap
